% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICHAELSR-package.R
\docType{package}
\name{MICHAELSR-package}
\alias{MICHAELSR}
\alias{MICHAELSR-package}
\title{MICHAELSR: The Title of the Project}
\description{
A paragraph providing a full description of the project (on several lines...)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SMAHSMAH/MICHAELSR}
  \item Report bugs at \url{https://github.com/SMAHSMAH/MICHAELSR/issues}
}

}
\author{
\strong{Maintainer}: Michael Smah \email{michael.smah@warwick.ac.uk} (\href{https://orcid.org/0000-0002-1062-2829}{ORCID}) [copyright holder]

}
\keyword{internal}
